# Allow different Ed25519/Curve25519 math versions