--- a/wolfssl/wolfcrypt/ge_operations.h
+++ b/wolfssl/wolfcrypt/ge_operations.h
@@ -29,9 +29,6 @@
 
 #ifdef HAVE_ED25519
 
-#if defined(HAVE___UINT128_T)
-    #include <stdint.h>
-#endif
 #include <wolfssl/wolfcrypt/fe_operations.h>
 
 /*