… configure.ac
… examples/client/client.c
… wolfcrypt/src/ge_operations.c
… wolfssl/internal.h
… wolfssl/wolfcrypt/settings.h