@dgarske 2017-06-21
@toddouska : No diference in any math logic, so I did not expect any performance change. The benchmarks below support that.

ED/CURVE Benchmark

NEW CODE:
```
./configure --enable-ed25519=small --enable-curve25519=small
CURVE 25519 key gen       270 ops took 1.003 sec, avg 3.713 ms, 269.315 ops/sec
CURVE 25519 key agree     300 ops took 1.116 sec, avg 3.721 ms, 268.746 ops/sec
ED    25519 key gen       263 ops took 1.000 sec, avg 3.803 ms, 262.917 ops/sec
ED    25519 sign          300 ops took 1.173 sec, avg 3.909 ms, 255.842 ops/sec
ED    25519 verify        200 ops took 1.583 sec, avg 7.917 ms, 126.311 ops/sec

./configure --enable-ed25519=small --enable-curve25519
CURVE 25519 key gen     13186 ops took 1.000 sec, avg 0.076 ms, 13185.525 ops/sec
CURVE 25519 key agree   13300 ops took 1.003 sec, avg 0.075 ms, 13264.649 ops/sec
ED    25519 key gen       262 ops took 1.002 sec, avg 3.823 ms, 261.578 ops/sec
ED    25519 sign          300 ops took 1.188 sec, avg 3.960 ms, 252.537 ops/sec
ED    25519 verify        200 ops took 1.563 sec, avg 7.817 ms, 127.932 ops/sec

./configure --enable-ed25519 --enable-curve25519=small
CURVE 25519 key gen       273 ops took 1.003 sec, avg 3.673 ms, 272.282 ops/sec
CURVE 25519 key agree     300 ops took 1.131 sec, avg 3.771 ms, 265.189 ops/sec
ED    25519 key gen     32643 ops took 1.000 sec, avg 0.031 ms, 32642.121 ops/sec
ED    25519 sign        22000 ops took 1.003 sec, avg 0.046 ms, 21937.503 ops/sec
ED    25519 verify      10700 ops took 1.005 sec, avg 0.094 ms, 10643.896 ops/sec

./configure --enable-ed25519 --enable-curve25519
CURVE 25519 key gen     12755 ops took 1.000 sec, avg 0.078 ms, 12754.629 ops/sec
CURVE 25519 key agree   12600 ops took 1.007 sec, avg 0.080 ms, 12514.230 ops/sec
ED    25519 key gen     32833 ops took 1.000 sec, avg 0.030 ms, 32832.413 ops/sec
ED    25519 sign        23100 ops took 1.006 sec, avg 0.044 ms, 22968.731 ops/sec
ED    25519 verify      11200 ops took 1.005 sec, avg 0.090 ms, 11139.735 ops/sec
```

CURRENT MASTER:
```
./configure --enable-ed25519=small --enable-curve25519=small
CURVE 25519 key gen       262 ops took 1.000 sec, avg 3.818 ms, 261.899 ops/sec
CURVE 25519 key agree     300 ops took 1.118 sec, avg 3.728 ms, 268.257 ops/sec
ED    25519 key gen       256 ops took 1.001 sec, avg 3.909 ms, 255.843 ops/sec
ED    25519 sign          300 ops took 1.184 sec, avg 3.946 ms, 253.401 ops/sec
ED    25519 verify        200 ops took 1.628 sec, avg 8.139 ms, 122.861 ops/sec

./configure --enable-ed25519 --enable-curve25519
CURVE 25519 key gen     12783 ops took 1.000 sec, avg 0.078 ms, 12782.451 ops/sec
CURVE 25519 key agree   13000 ops took 1.004 sec, avg 0.077 ms, 12948.257 ops/sec
ED    25519 key gen     33749 ops took 1.000 sec, avg 0.030 ms, 33748.397 ops/sec
ED    25519 sign        28900 ops took 1.000 sec, avg 0.035 ms, 28897.396 ops/sec
ED    25519 verify      10600 ops took 1.004 sec, avg 0.095 ms, 10556.654 ops/sec
```